requirements.txt
requests
tqdm
pandas
orjson
python-dotenv
pre-commit
# later (optional)
pyserini
networkx
sentence-transformers
faiss-cpu
